kadarbitrmcp/pyproject.toml
[project]
name = "arbitr-mcp"
version = "0.1.0"
description = "MCP server for kad.arbitr.ru via api-assist.com"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "prometheus-client>=0.20.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["tools"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["test"]


